Complete consolidation of the label views
